refactor: Rename cesium query & popup services
