+++ projects/hslayers-cesium/src/query-popup.service.ts
